+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - PR #32 - Enable filter coefficient reuse across multiple calls to resample_poly and performance bug fixes
 - PR #34 - Implement CuPy kernels as module object and templating
 - PR #35 - Make upfirdn's kernel caching more generic; support 2D
+- PR #36 - Set default upfirdn/resample_poly behavior to use Raw CuPy CUDA kernel rather than Numba; Doc updates
 
 ## Bug Fixes
 - PR #4 - Direct method convolution isn't supported in CuPy, defaulting to NumPy [Examine in future for performance]